chore: move to pnpm
The main motivation being that with pnpm packages
in the monorepo don't have to specify versions of
the dependencies of other packages in the monorepo.
It is also widely adopted across different monorepos.

As well as `changesets` would replace the `workspace:*`
marked versions on fly, so there won't be a need in managing
those by hands (if there was one before).

Fixes the lint issue, where in `with-next-auth`
example a lint config was not properly set up, which caused the `lint`
job to hang up waiting for stdin input.

Adds a `turbo.json` config to `auth-kit` to depend on `auth-client`,
as the build job for the first depends on the latter.