feat!: move to `react-query`
It is a breaking change, since the hook APIs also change.
However, everything is much more reliable now, and effects are not triggered
when they don't have to.

The context was changed to simply hold the `config`, which one has now
to initiate with `createConfig`, in order to construct an `appClient`.

The rest of the state now is handled by `zustand`, specifically, the
`useProfile` and `useSignInMessage` use it.